Rename vault folder to org-vault
