+++ apps/web/src/app/organizations/org-vault/collection-badge/collection-name.badge.component.ts
