+++ apps/web/src/app/organizations/org-vault/group-badge/group-name-badge.component.html
