+++ apps/web/src/vault/app/vault/vault.module.ts
@@ -2,8 +2,8 @@ import { NgModule } from "@angular/core";
 
 import { BreadcrumbsModule } from "@bitwarden/components";
 
-import { CollectionBadgeModule } from "../../../app/organizations/vault/collection-badge/collection-badge.module";
-import { GroupBadgeModule } from "../../../app/organizations/vault/group-badge/group-badge.module";
+import { CollectionBadgeModule } from "../../../app/organizations/org-vault/collection-badge/collection-badge.module";
+import { GroupBadgeModule } from "../../../app/organizations/org-vault/group-badge/group-badge.module";
 import { LooseComponentsModule, SharedModule } from "../../../app/shared";
 
 import { BulkDialogsModule } from "./bulk-action-dialogs/bulk-dialogs.module";
